Add code to manipulate display
I first added code to dynamically change the display, then the clear
function, so that pressing the "C" button would revert the display to 0.
Then, I added a class to each number, selected them all, and added an
event listener to pass each value to the display, though only one at a
time.