Code single-digit arithemtic functions
The calculator will now display and perform single-digit calculations.
This is not a valid solution, but merely the first step. The current
functionality relies on a user clicking buttons in a very specific
order. They MUST click a number, an operator, a number, and then...
Well, ideally the equals sign, but anything at all will trigger the
arithemtic.

I used a simple counter (count) to keep track of which input is which.
The first input (count = 1) is the first number. The second input
(count = 2) is the operator, the third input (count = 3) is the
second number, and the fourth input (count = 4) is the arithemtic
function call, and then the counter is reset to 0.

This does not take into account the user clicking anything other than
what is expected each time. It also does not allow the user to enter
multi-digit numbers, like "11", or anything with a decimal point. It
does not account for division by 0. These are all things I must figure
out going forward.