Edit code to allow for multi-digit numbers
I initially had all of the keys classed as "button", then was listening
for any of them to be clicked, and trying to differentiate which was
which. It was kind of working, but with a ton of nested if-else
statements to try to account for every scenario. I went in and changed
the numbers to class "numeral" and the operators to class "operator",
and was then able to listen JUST for either of them. I kept the counter
variable, as that lets the script know which entry is being stored.
When counter = 1, it's the first entry, when counter = 2, it's the
second entry. I added logic to account for when to call operate(); if
there's no secondNumber, we just store the operator. If secondNumber
exists, then we call the function, the result becomes firstNumber,
and secondNumber is cleared. This way, the next operator selected will
wait for another number to be entered and a new operator.

I still have a couple of things to change. If an operator is pressed
before a number, then 0 should be passed to firstNumber. This should be
as simple as adding an if statement to the operator listener.

There's also the matter of repeated functions - if the equals key is
pressed, the operate() function returns the answer, eg x + y, but if
the equals key is pressed again, the function should be called again,
this time as (x + y) + y.

And, of course, I need to get the % and +/- keys working. I have
functions written for them, it's just a matter of getting them to work.