Add code to finalize multi-digit calculations
Created a separate class for the equals sign, as I the functionality
needed to be a little different than the other operators, and this
seemed to make more sense than a bevy of nested if-else statements.
The calculator can now handle no first number, no second number, an
operator clicked twice, and the equals sign clicked twice.

I still need to handle the case of equals being pressed before an
operator, and then the percent and +/- keys.