Continue to refine basic maths
Added a variable endOfCalc that helps handle whether the expressions
are to continue or not. When it is 0, the calculation keeps going, when
it is 1, the display should clear and calculations start over. Where
this gave me a bit of trouble was clicking an operator after clicking
equals. This should "restart" the chain, so I had to wipe the
secondNumber variable, set firstNumber to the display, but also evade
the displayClear I'd built into the numerals listener. Turns out, I
just needed to set endOfCalc back to 0.