Add code to handle decmials
I thought this was working already, as I couldn't enter "1.32.45.5",
for example, but then I realized that trying to enter a second decimal
just caused the script to stop reading in the number. So, I changed the
class of the decimal key to be unique, and added a listener to it, so
that when it is pressed, it inserts the decimal just once, and then
ignores it.