Adjust decimal handling and display
Decimal wasn't displaying correctly. If the user wanted to enter "0.3",
nothing would show up until they entered the 3. Similarly, there was
no indication of the user entered a decimal in the middle of a number.

Still trying to get the trailing 0 to show up while typing (e.g. if the
number entered is 3.3004, the display currently just shows 3.3 until the
4 is entered, instead of 3.30 and 3.300), and not display a leading
zero if a user types a zero at the start of their entry.