Edit script to handle zeroes
I had been converting the entries to floats to display them, as that
took care of inadvertant leading zeroes. However, that also prevented
the display from showing intermittent zeroes in decimals. Instead, I've
left the strings as strings, and added in a check that if the string is
longer than 1 and starts with 0, slice off that zero and store that
value as the entry. This removes leading zeroes, displays intermittent
zeroes in floats, and displays correctly even with decimals.