Add code to limit length of display
I added an updateDisplay function that could have an if statement and
truncate the length of the display if it went over a certain number. I
don't LOVE this solution, as it still has some issues. For one, it
doesn't stop the entry, just stops showing it. Another issue, and this
is really just my own aesthetic, but exponential numbers don't display
quite right, in that the "LEDs" in the background no longer line up.
Functionally, though, everything works.