Fix background LCD numbers
I've been calling them LEDs, but they're not, they're supposed to
simulate LCD numbers. Either way, I think I found a better way to
position them so that they stay where they should be, regardless of
screen size. Instead of trying to "dial in" the top and right settings,
I wrapped the background in a div that is a child of the display text.
Then, absolute is referencing the display text itself. Setting the
div to inheret the width of the parent div forces it to align correctly.